+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 ## 1.4.0
 
-- Increased system minimum version to 4.3
+- Increased system minimum version to 4.3 & marked v12 maximum
 - Removed ability for users with Limited permissions to view an Organization sheet
 - Updated uses of Dialog to DialogV2
 - Adjusted code to account for system deprecations
-- Added Kingdoms & Warfare to registered source books
+- Added Kingdoms & Warfare to registered source books as "K&W"
 - Added labels to data model fields
 
 ## 1.3.0